Dirty fix to only save tweets with #foafu from stream
statuses/filter doesn't combine the users and track filters which causes the system to save all tweets by either one of the users or tweets with #foafu but we need tweets bu both one of the users AND the #foafu hashtag. This is now fixed quick and dirty with an details.text indexOf check.